addressing some of eli's refactoring suggestions
