+++ chirho/observational/handlers/soft_conditioning.py
@@ -1,24 +1,10 @@
 import functools
 import operator
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    Literal,
-    Optional,
-    Protocol,
-    Set,
-    TypedDict,
-    TypeVar,
-    Union,
-)
+from typing import Callable, Literal, Optional, Protocol, TypedDict, TypeVar, Union
 
 import pyro
 import pyro.distributions.constraints as constraints
 import torch
-
-from chirho.indexed.handlers import add_indices
-from chirho.indexed.ops import IndexSet, gather, indexset_as_mask, scatter
 
 T = TypeVar("T")
 
@@ -204,127 +190,3 @@ class AutoSoftConditioning(pyro.infer.reparam.strategies.Strategy):
         raise NotImplementedError(
             f"Could not reparameterize deterministic site {msg['name']}"
         )
-
-
-class CutModule(pyro.poutine.messenger.Messenger):
-    """
-    Converts a Pyro model into a module using the "cut" operation
-    """
-
-    vars: Set[str]
-
-    def __init__(self, vars: Set[str]):
-        self.vars = vars
-        super().__init__()
-
-    def _pyro_sample(self, msg: Dict[str, Any]) -> None:
-        # There are 4 cases to consider for a sample site:
-        # 1. The site appears in self.vars and is observed
-        # 2. The site appears in self.vars and is not observed
-        # 3. The site does not appear in self.vars and is observed
-        # 4. The site does not appear in self.vars and is not observed
-        if msg["name"] not in self.vars:
-            if msg["is_observed"]:
-                # use mask to remove the contribution of this observed site to the model log-joint
-                msg["mask"] = (
-                    msg["mask"] if msg["mask"] is not None else True
-                ) & torch.tensor(False, dtype=torch.bool).expand(msg["fn"].batch_shape)
-            else:
-                pass
-
-        # For sites that do not appear in module, rename them to avoid naming conflict
-        if msg["name"] not in self.vars:
-            msg["name"] = f"{msg['name']}_nuisance"
-
-
-class CutComplementModule(pyro.poutine.messenger.Messenger):
-    vars: Set[str]
-
-    def __init__(self, vars: Set[str]):
-        self.vars = vars
-        super().__init__()
-
-    def _pyro_sample(self, msg: Dict[str, Any]) -> None:
-        # There are 4 cases to consider for a sample site:
-        # 1. The site appears in self.vars and is observed
-        # 2. The site appears in self.vars and is not observed
-        # 3. The site does not appear in self.vars and is observed
-        # 4. The site does not appear in self.vars and is not observed
-        if msg["name"] in self.vars:
-            # use mask to remove the contribution of this observed site to the model log-joint
-            msg["mask"] = (
-                msg["mask"] if msg["mask"] is not None else True
-            ) & torch.tensor(False, dtype=torch.bool).expand(msg["fn"].batch_shape)
-
-    def _pyro_post_sample(self, msg: Dict[str, Any]) -> None:
-        if msg["name"] in self.vars:
-            assert msg["is_observed"]
-
-
-def cut(model: Optional[Callable] = None, *, vars: Set[str] = set()):
-    if model is None:
-        return functools.partial(cut, vars=vars)
-
-    return CutModule(vars)(model), CutComplementModule(vars)(model)
-
-
-class IndexCutModule(
-    pyro.poutine.messenger.Messenger
-):  # TODO subclass DependentMaskMessenger
-    """
-    Represent module and complement in a single Pyro model using plates
-    """
-
-    vars: Set[str]
-    name: str
-
-    def __init__(self, vars: Set[str], *, name: str = "__cut_plate"):
-        self.vars = vars
-        self.name = name
-        super().__init__()
-
-    def __enter__(self):
-        add_indices(IndexSet(**{self.name: {0, 1}}))
-        return super().__enter__()
-
-    def _pyro_sample(self, msg: Dict[str, Any]) -> None:
-        if pyro.poutine.util.site_is_subsample(msg):
-            return
-        # There are 4 cases to consider for a sample site:
-        # 1. The site appears in self.vars and is observed
-        # 2. The site appears in self.vars and is not observed
-        # 3. The site does not appear in self.vars and is observed
-        # 4. The site does not appear in self.vars and is not observed
-
-        # TODO inherit this logic from indexed.handlers.DependentMaskMessenger
-        # use mask to remove the contribution of this observed site to the model log-joint
-
-        cut_index = IndexSet(**{self.name: {0 if msg["name"] in self.vars else 1}})
-        mask = indexset_as_mask(cut_index)  # TODO device
-        msg["mask"] = mask if msg["mask"] is None else msg["mask"] & mask
-
-        # expand distribution to make sure two copies of a variable are sampled
-        msg["fn"] = msg["fn"].expand(
-            torch.broadcast_shapes(msg["fn"].batch_shape, mask.shape)
-        )
-
-    def _pyro_post_sample(self, msg: Dict[str, Any]) -> None:
-        if pyro.poutine.util.site_is_subsample(msg):
-            return
-
-        if (not msg["is_observed"]) and (msg["name"] in self.vars):
-            # discard the second value
-            # TODO is the mask computed in _pyro_sample correct?
-            value_one = gather(
-                msg["value"],
-                IndexSet(**{self.name: {0}}),
-                event_dim=msg["fn"].event_dim,
-            )
-
-            msg["value"] = scatter(
-                {
-                    IndexSet(**{self.name: {0}}): value_one,
-                    IndexSet(**{self.name: {1}}): value_one.detach(),
-                },
-                event_dim=msg["fn"].event_dim,
-            )
